- Shoot through crosshair
- line trace through crosshair

we got the screen space location of our crosshairs D projected it to WorldSpace and

then performed a line Trace straight outward from the World Space location of our crosshairs.

We then use the hit result from that line trace to set our beam and location and our impact location for the impact particles.